requirements.txt
hydra-zen
unique-names-generator
torchmetrics
plotext
tqdm
wandb
isort
black
autoflake
pre-commit
